+++ src/main/scala/org/constellation/serialization/ConstellationKryoRegistrar.scala
@@ -9,7 +9,6 @@ import org.constellation.gossip.state.GossipMessage
 import org.constellation.infrastructure.endpoints.BuildInfoEndpoints.BuildInfoJson
 import org.constellation.rewards.EigenTrustAgents
 import org.constellation.schema.serialization.ExplicitKryoRegistrar
-import org.constellation.schema.snapshot.SnapshotProposalPayload
 
 object ConstellationKryoRegistrar
     extends ExplicitKryoRegistrar(
@@ -23,6 +22,6 @@ object ConstellationKryoRegistrar
         (classOf[StoredRewards], 185),
         (classOf[BuildInfoJson], 187),
         (classOf[GossipPath], 1032),
-        (classOf[GossipMessage[SnapshotProposalPayload]], 1033),
+        (classOf[GossipMessage[_]], 1033)
       )
     ) {}